refactor(export): move wiki export utility from agents/ to core/
ExportAgent makes no LLM calls and has no LLMProvider dependency — it is
a pure serialisation utility and does not belong in the agents package.

Changes:
- Move synthadoc/agents/export_agent.py -> synthadoc/core/export.py
- Remove artificial run()/_run() split (single run() method is enough
  with no BaseAgent error-suppression wrapper involved)
- Remove ExportAgent from agents/_base.py exemption note; update to
  reference the new location generically
- Update all import paths: mcp_server.py, http_server.py
- Rename tests/test_export_agent.py -> tests/test_export.py; update
  all import paths in test files
- Update test_base_agent.py: test now asserts ExportAgent lives in
  core/ and has no BaseAgent dependency (not an exemption edge case)
- Update 3 patch targets in test_mcp_server.py
- docs/design.md: remove duplicate ExportAgent subsection from §4
  Agents; update §24 Export Formats intro to clarify it is a pure
  serialisation utility with no LLM dependency
- badges.json: agents count 12 -> 11 (ExportAgent no longer scanned
  by count_agents() in agents/ directory)
