+++ docs/badges.json
@@ -6,6 +6,6 @@
   "coverage": 91,
   "coverage_color": "brightgreen",
   "mcp_tools": 12,
-  "agents": 12,
+  "agents": 11,
   "maintenance_workflows": 5
 }
