+++ synthadoc/agents/_base.py
@@ -6,9 +6,9 @@ All agents that make direct LLM calls without a tool-calling loop must
 inherit from ``BaseAgent``.  Agents that run an agentic tool loop inherit
 from ``AgenticWorkflow`` instead (see ``agents/workflows/_base.py``).
 
-Non-LLM utility classes (``ExportAgent``) are exempt: they follow the
-``run()`` / ``_run()`` naming convention but do not inherit ``BaseAgent``
-because they require no LLM provider.
+Non-LLM utility classes (e.g. the wiki export serialiser in
+``core/export.py``) live outside the agents package and do not inherit
+``BaseAgent`` because they require no LLM provider.
 
 Contract
 --------
