+++ synthadoc/integration/mcp_server.py
@@ -1,4 +1,4 @@
-# SPDX-License-Identifier: AGPL-3.0-or-later
+﻿# SPDX-License-Identifier: AGPL-3.0-or-later
 # Copyright (C) 2026 Paul Chen / axoviq.com
 from __future__ import annotations
 from pathlib import Path
@@ -71,7 +71,7 @@ def create_mcp_server(orchestrator):
         Other formats without output_path: {"format", "content": str, "pages": N}
         """
         from datetime import date
-        from synthadoc.agents.export_agent import ExportAgent, ExportOptions, EXPORT_FORMATS
+        from synthadoc.core.export import ExportAgent, ExportOptions, EXPORT_FORMATS
         if format not in EXPORT_FORMATS:
             return {"error": f"unknown format {format!r}. Valid: {sorted(EXPORT_FORMATS)}"}
